requirements.txt
streamlit
PyPDF2
pillow
pytesseract
langchain
langchain-community
faiss-cpu
groq
transformers
torch
tqdm